Add OpenAI conversation entity
