+++ homeassistant/components/openai_conversation/const.py
@@ -1,6 +1,9 @@
 """Constants for the OpenAI Conversation integration."""
 
+import logging
+
 DOMAIN = "openai_conversation"
+LOGGER = logging.getLogger(__name__)
 CONF_PROMPT = "prompt"
 DEFAULT_PROMPT = """This smart home is controlled by Home Assistant.
